Rewrite zlib cross-validation tests with compile-once architecture
Replace ~50 individual test compilations with two reusable Wado driver
programs (compress + inflate) compiled once via OnceLock<Component>.
Each test now only performs cheap Wasm instantiation, eliminating memory
pressure from concurrent compilations that caused SIGSEGV flakiness.

- Add fixtures/sub/zlib_cross_{compress,inflate}.wado stdin-based drivers
- Rename zlib_rs.rs → zlib_interop.rs (focus is interoperability)
- Add fuzz tests with deterministic PRNG for zlib/raw/gzip round-trips
- All 49 tests pass in ~16s (was flaky with concurrent compilations)
